requirements.txt
fastapi==0.120.2
uvicorn==0.38.0

# Database
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
asyncpg==0.30.0
alembic==1.17.1

# Auth / Security
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.1.3
email-validator==2.3.0

# config
python-dotenv==1.2.1
pydantic==2.12.3
pydantic-settings==2.11.0

# HTTP
httpx==0.28.1
requests==2.32.5

# Redis
redis==5.0.4

# MinIO
minio==7.2.9

# Qdrant
qdrant-client==1.8.1

# Utils
loguru==0.7.3
orjson==3.11.4

#Tests
pytest
pytest-asyncio

# LLM
openai
openai-agents
guardrails-ai


# parser
fake_useragent
beautifulsoup4
lxml
aiohttp

# ocr
google-cloud-vision
torch
transformers
pillow
